third_party/WebKit/LayoutTests/imported/web-platform-tests/dom/lists/DOMTokenList-value-expected.txt

This is a testharness.js-based test.
FAIL DOMTokenList value assert_equals: length should be the number of tokens expected 2 but got 3
Harness: the test ran to completion.
